Code Generation

This is the backend of Comonicon. Currently it can only generates a Julia AST.

References

Comonicon.CodeGen.ASTCtx — Type

ASTCtx

AST code generation context.

Comonicon.CodeGen.codegen — Function

codegen(ctx, cmd)

Generate target code according to given context ctx from a command object cmd.

Comonicon.CodeGen.codegen — Method

codegen(cmd)

Generate Julia AST from given command object cmd. This will wrap all the generated AST in a function command_main.

Comonicon.CodeGen.prettify — Method

prettify(ex)

Prettify given expression, remove all LineNumberNode and extra code blocks.

Comonicon.CodeGen.pushmaybe! — Method

push!(args, item)

Push an item to the expression or list. Do nothing if the item is nothing.

Comonicon.CodeGen.read_arg — Method

read_arg(ctx::ASTCtx)

Return the expression that read the next argument from ARGS.

Comonicon.CodeGen.rm_lineinfo — Method

rm_lineinfo(ex)

Remove LineNumberNode in a given expression

Comonicon.CodeGen.xerror — Method

xerror(msg)

Create an expression that contains error message, automatically merge interpolation expressions.

Comonicon.CodeGen.xmatch — Function

xmatch(regexes, actions, str[, st = 1])

Generate a long ifelse expression that acts like a pattern matching expression that match given regex list and do the corresponding actions.