Build and Install CLI

You can use various way to build and install your CLI, this includes:

use it as a script, and enable or disable compile cache via (Comonicon.disable_cache and Comonicon.enable_cache).

build it as a package and install to ~/.julia/bin:

use compile=:min in Comonicon.install if you don't care about the speed

use sysimg=true in Comonicon.install if you care about both start up time and the speed

Reference

Comonicon.Comonicon — Module

All the terminals are under my command. Comonicon is a CLI (Command Line Interface) generator that features light-weight dependency (optional to have zero dependency), fast start-up time and easy to use. See the website for more info.

Comonicon.PATH — Module

path related functions.

Comonicon.PATH.default_exename — Method

default_exename()

Default Julia executable name: joinpath(Sys.BINDIR, Base.julia_exename())

Comonicon.PATH.default_julia_bin — Method

default_julia_bin()

Return the default path to .julia/bin.

Comonicon.PATH.default_julia_fpath — Method

default_julia_fpath()

Return the default path to .julia/completions