Build and Install CLI

You can use various way to build and install your CLI, this includes:

use it as a script, and enable or disable compile cache via (Comonicon.disable_cache and Comonicon.enable_cache).

build it as a package and install to ~/.julia/bin:

use compile=:min in Comonicon.install if you don't care about the speed

use sysimg=true in Comonicon.install if you care about both start up time and the speed

Reference

Comonicon.Comonicon — Module

All the terminals are under my command. Comonicon is a CLI (Command Line Interface) generator that features light-weight dependency (optional to have zero dependency), fast start-up time and easy to use. See the website for more info.

Base.write — Method

write([io], cmd::EntryCommand)

Write the generated CLI script into a Julia script file. Default is the cachefile.

Comonicon.build — Function

build(mod[, name=default_name(mod)]; kwargs...)

Build system image for given CLI module. See also install.

Comonicon.build — Method

build()

Build a system image for Comonicon.

Comonicon.cmd_script — Method

cmd_script(mod, shadow; kwargs...)

Generates a shell script that can be use as the entry of mod.command_main.

Arguments

mod: a module that contains the commands and the entry.

shadow: location of a Julia script that calls the actual mod.command_main.

Keywords

exename: The julia executable name, default is PATH.default_exename.

sysimg: System image to use, default is PATH.default_sysimg.

project: the project path of the CLI.

compile: julia compile level, can be [:yes, :no, :all, :min]

optimize: julia optimization level, default is 2.

Comonicon.install — Function

install(mod[, name=default_name(mod)]; kwargs...)

Install the CLI defined in module mod to the bin.

Arguments

mod: module that contains the CLI, usually a project module.

name: default is the default_name.

Keywords

bin: path of the bin folder, default is the ~/.julia/bin.

exename: Julia executable name, default is PATH.default_exename.

project: the project path of the CLI.

compile: julia compile level, can be [:yes, :no, :all, :min]

optimize: julia optimization level, default is 2.

sysimg: compile a system image using PackageCompiler or not, default is false.

incremental: PackageCompiler option, default is false.

filter_stdlibs: PackageCompiler option, default is false.

cpu_target: PackageCompiler option. default is "native".

lib_path: the lib folder that stores the compiled system image, default is the <project>/deps/lib.

Comonicon.precompile_script — Method

precompile_script(mod)

Generates a script to execute as precompile_execution_file for all the commands.

Comonicon.PATH.default_exename — Method

default_exename()

Default Julia executable name: joinpath(Sys.BINDIR, Base.julia_exename())

Comonicon.PATH.default_sysimg — Method

default_sysimg()

Return libcomonicon.(dylib, so, dll) if the package Comonicon is compile by PackageCompiler. Or return nothing.