Build and Install Tools

You can use various way to build and install your CLI, this includes:

use it as a script, and enable or disable compile cache via (Comonicon.disable_cache and Comonicon.enable_cache).

build it as a package and install to ~/.julia/bin:

use compile=:min in Comonicon.install if you don't care about the speed

use sysimg=true in Comonicon.install if you care about both start up time and the speed

Reference

ComoniconBuilder.cmd_script — Method

cmd_script(mod, shadow; kwargs...)

Generates a shell script that can be use as the entry of mod.command_main.

Arguments

mod: a module that contains the commands and the entry.

shadow: location of a Julia script that calls the actual mod.command_main.

Keywords

exename: The julia executable name, default is default_exename.

sysimg: System image to use, default is nothing.

project: the project path of the CLI.

compile: julia compile level, can be [:yes, :no, :all, :min]

optimize: julia optimization level, default is 2.

ComoniconBuilder.default_exename — Method

default_exename()

Default Julia executable name: joinpath(Sys.BINDIR, Base.julia_exename())

ComoniconBuilder.default_julia_bin — Method

default_julia_bin()

Return the default path to .julia/bin.

ComoniconBuilder.default_julia_fpath — Method

default_julia_fpath()

Return the default path to .julia/completions

ComoniconBuilder.detect_shell — Method

detect_shell()

Detect shell type via SHELL environment variable.

ComoniconBuilder.install_script — Method

install a script as the CLI

ComoniconBuilder.osname — Method

osname()

Return the name of OS, will be used in building tarball.