Parse

This is the frontend @cast and @main of Comonicon.

The @main command will use generate a few functions in the module:

the entry function for CLI command_main.

comonicon_install: for command build and installation.

comonicon_install_path: for path build and installation.

julia_main: for building standalone applications.

References

Comonicon.Comonicon — Module

All the terminals are under my command. Comonicon is a CLI (Command Line Interface) generator that features light-weight dependency (optional to have zero dependency), fast start-up time and easy to use. See the website for more info.

Comonicon.default_name — Method

default_name(x::String)

Return the lowercase of a given package name. It will ignore the suffix if it ends with ".jl".

Comonicon.get_version — Method

get_version(m::Module)

Get the version of a given module. It will try to find the version of Project.toml if the given module is a project module. If fails, it returns v"0.0.0".

Comonicon.rm_format — Function

rm_format(md)

Remove Markdown DOM and flatten to strings.

Comonicon.set_cmd! — Function

set_cmd!(cmds::Dict, cmd)

register cmd in the command registry cmds, which usually is a constant CASTED_COMMANDS under given module.