Command Types

The frontend @main and @cast restrict a lot features to make the package interface as simple as possible. However, if you want to customize the behaviour, you can also create command objects directly via the interfaces at this level, which are all in Comonicon.AST.

References

Comonicon.AST.brief — Method

brief(text::String)

Use the first sentence as the brief description.

Comonicon.AST.splitlines — Function

splitlines(s, width = 80)

Split a given string into lines of width 80 characters.

Comonicon.AST.splittext — Method

splittext(s)

Split the text in string s into an array, but keep all the separators attached to the preceding word.

Note

this is copied from Luxor/text.jl